backend/requirements.txt
Flask
flask_cors
langchain_community
protobuf
python-dotenv
google-generativeai
